requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
requests
python-dotenv
whatchdog